src/Python/requirements.txt
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
certifi==2024.12.14
cffi==1.15.1
charset-normalizer==3.4.0
cryptography==44.0.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.8.0
requests==2.31.0
six==1.17.0
typing_extensions==4.7.1
urllib3==2.0.7
questionary==1.10.0
readchar==4.2.0
